% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVarLabelTable.R
\name{getVarLabelTable}
\alias{getVarLabelTable}
\title{getVarLabelTable}
\usage{
getVarLabelTable(SAS.SPSS_Data, import.method = "foreign")
}
\arguments{
\item{SAS.SPSS_Data}{a data.frame object imported via 'foreign' or 'Hmisc'}

\item{import.method}{the import method used: foreign or Hmisc}
}
\value{
A data.frame object containing the names of the data.frame, along with the corresponding variable labels.
}
\description{
getvarLabelTable will extract the variable names and variable labels from a data.frame imported from an SPSS or SAS file via either the foreign or Hmisc packages
}
\examples{
#NULL
}

