% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.packagesArchive.R
\name{install.packagesArchive}
\alias{install.packagesArchive}
\title{install.packagesArchive}
\usage{
install.packagesArchive(package, version)
}
\arguments{
\item{package}{Name of the package}

\item{version}{Version of the package}
}
\description{
install.packagesArchive in a convenience wrapper for installing archived versions of packages from CRAN.
}
\examples{
#install.packagesArchive(package = "mixlow", version = "1.0.1")
#install.packagesArchive(package = "mixlow", version = "2.0.1")
}

