% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_workbook.R
\name{read_workbook}
\alias{read_workbook}
\title{read_workbook}
\usage{
read_workbook(ExcelFile, na_value = "", add_to_Envir = TRUE,
  output_ls = FALSE)
}
\arguments{
\item{ExcelFile}{Character string of the directory and filename pointing at the file to load}

\item{na_value}{Value in the excel workbook that should be treated as missing/NA}

\item{add_to_Envir}{TRUE/FALSE: Should each sheet be added to the global environment, each as a data.frame}

\item{output_ls}{TRUE/FALSE: Should each sheet be appended to a list of data.frames}
}
\value{
if output_ls==TRUE, a list of data.frames.
}
\description{
read_workbook is an extension of readxl::read_excel, loading all sheets in an excel workbook either directly into the global environment or as a part of a list
}
\examples{
#NULL
}

