% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceDir.R
\name{sourceDir}
\alias{sourceDir}
\title{sourceDir}
\usage{
sourceDir(path, verbose = TRUE)
}
\arguments{
\item{path}{Path to folder containing all files to source.}

\item{verbose}{If TRUE, prints the name of each file sourced to the console}
}
\description{
sourceDir sources all .R files contained at a given path directory
}
\examples{
#NULL
}

