% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_github_pkgs.R
\name{update_github_pkgs}
\alias{update_github_pkgs}
\title{update_github_pkgs}
\usage{
update_github_pkgs()
}
\value{
N/A
}
\description{
update_github_pkgs updates all github packages
}
\examples{
update_github_pkgs()
}

